Protype-S-ai/requirements.txt

# Protype.AI - Developed by Islam Ibrahim, Carrot Studio
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Database
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9

# AI & NLP
spacy==3.7.2
scikit-learn==1.3.2
google-generativeai==0.3.1
transformers==4.35.2

# Search
elasticsearch==8.11.0
serpapi==0.1.0

# Background processing
celery==5.3.6
redis==5.0.1

# Web scraping & data processing
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
numpy==1.26.2

# Visualization
networkx==3.2.1
matplotlib==3.8.2
plotly==5.18.0

# Text-to-speech
pydub==0.25.1
elevenlabs==0.2.26


# JavaScript build tools (for React frontend)
# Note: These are installed via npm, not pip
# npm install -g create-react-app
# ✅ مكتبات الذكاء الاصطناعي والتعلم العميق (Deep Learning & Transformers)
torch==2.0.1+cpu  # PyTorch - الإصدار المناسب لـ CPU
transformers>=4.38.1  # أحدث إصدار لدعم نماذج DeepSeek وLLaMA
torch-geometric==2.3.0  # لدعم Graph Neural Networks (GNN)

nltk>=3.8.1  # مكتبة تحليل اللغة الطبيعية
sentence-transformers>=2.2.2  # استخراج ميزات النصوص باستخدام Transformers


# ✅ دعم الإدراك الزمني والتعامل مع التواريخ
pytz>=2024.1  # إدارة المناطق الزمنية
dateutil>=2.8.2  # تحليل التواريخ والتوقيت


# ✅ مكتبات اختيارية لتحليل الصور والصوت (إذا احتجت إليها لاحقًا)
torchvision>=0.15.0+cpu  # معالجة الصور
pillow>=10.2.0  # دعم الصور في Python
torchaudio>=2.0.0+cpu  # معالجة الصوت